Rename IDENTIFYING AND MITIGATING CYBER RISK to IDENTIFYING AND MITIGATING CYBER RISK.
Cybersecurity has become a critical aspect of our daily lives as we rely more and more on technology and the internet to communicate, work, and access information. With the ever-evolving nature of cyber threats, it’s crucial to stay informed and prepared to defend against them. In this article, we’ll explore current cyber threats, social engineering tactics, and how to stay safe in the online world.


https://medium.com/@EberechiDan/cybersecurity-has-become-a-critical-aspect-of-our-daily-lives-as-we-rely-more-and-more-on-ba77e2b25455